fix: move to own file, abandon cupy
